Explicitly specify AZ for boot volume
If a volume is to be used as a boot volume (i.e. attached to the
Instance prior to creation using the Volumes property, rather than
afterwards with a VolumeAttachment), then it cannot obtain the
AvailabilityZone (which, if not specified, is assigned dynamically) from
the Instance, since that has not yet been created and assigned to an AZ.
Instead, require the AZ to be specified as a parameter.

Change-Id: Icccc9b531b1fd7754d76c13d90679ed04c9c0d02