Add sudo -E to heat-jeos recommended command
Normally heat would be run without root capabilities.  The
directory /var/lib/libvirt/images is typically not world
readable.  As a result, the glance operation requires root
permissions to actually read the file, and may be a source
of confusion for users.

Also break the notification into two lines since it wraps on
80 character windows.

Change-Id: I905f3c24ed253ddc1e843bc73c93464c65d3df0b