Adapt WordPress_NoKey to changes in str_replace
Adapt naming of parameters to the str_replace function to implementation
changes in that function so the replacement works correctly. The following
change introduces exact string matching, so we have to change from the
previous python string template based notation:

https://review.openstack.org/#/c/47772

Note: the change to the HOT template depends on the change above being
merged (or gets necessary as soon as this change is merged).

Change-Id: I229bee76771b0b0f691943f9fa8ceaa8393de781
Closes-Bug: #1229741