Remove AWS style syntax from HOT templates
This patch removed AWS style syntax from samples in the heat-templates
repository to not advocate their use in future HOT templates.

The current HOT template parser in the engine uses a camel case to snake case
conversion but this is going to be removed in favor of strict HOT keyword
checking. Therefore, we do not want to advocate syntax in HOT templates that
will break in the future.

Change-Id: I2aab31b64c863d2a62d3c7e57c0fb983419c3186