HOT template to deploy servers in new neutron net
Add a HOT template to create a new neutron network and router to the public net
and to deploy two servers into the new network. Also assign floating IP from
public network to the two servers.
There is no special assumption on the image used for deployment since no
configuration inside the instance is done. The template is really just for
showing the base neutron configuration.

Change-Id: I8e177c3df3c82b37a163b13ee34b7c914b560479