Example template that performs copying of SSH keys
This template shows how to leverage the new resources for software
configuration/deployment, particularly the use of scripts for
configuration, to enable copying of SSH keys in a similar fashion as
'ssh-copy-id'. The example makes use of configuration inputs, and
outputs, and get_attr. The use of get_attr introduces an implicit
dependencyOne example to show ssh-copy-id between servers

Change-Id: Ic15c325d1c6f35108206d0e6e8dc7339c1383a30