Add a script to validate templates
This adds a new script that can walk through a template directory and
call heat template-validate on them. It handles environment files by
expecting $template_env.ext, so it'll need renaming for the 2 templates
that we have needing that. It also needs #1298450 fixed. The goal is to
be able to use this script in a future gate.

Change-Id: I4246d494de9a3a56c3b334ba20ad06186ccaefd6