Remove misleading OS choices from legacy F17 examples
These examples all mention alternative OS versions, which is confusing
to users, especially the ubuntu one, which will definitely not work and
evidently is causing confusion for users.

So remove the references to anything other than F17 (which these did work
on at one point).  Obviously we still have to refresh these at some point
to make sure they still work on a non-EOL version of Fedora, but hopefully
this somewhat reduces confusion in the meantime.

Change-Id: I9d40f44f86e40f7d67fa4c3340b7fc6ad01bf270