HOT for IIS with Drupal
It deploys a new Drupal 7 web site on a server with IIS
and PHP 5.4 configured. Web site uses a local database on
SQL Server 2008 R2 SP2 Express

Required tools are downloaded from the internet using
official links. SHA1 hashes are computed for the downloaded files
and compared with the expected hashes to ensure file integrity.

The unit tests for the user data scripts are written using
Pester 3.0

Change-Id: Ia9c12db0a92fc4f39be9056c67471eb8c136a86a
Partially-Implements: blueprint windows-instances