heat-config-puppet: support hiera
Updates the puppet hook so that we support using Hiera in addition
to the existing Facter based input mechanism. Hiera provides
a more flexible way of dealing with puppet parameters and
seems to be prefered in the Puppet community as well.

Use of the Hiera data will also require an additional
'hiera' element which will probably live in the
triple-puppet-elements tree.

Change-Id: I497f2d2286025e48e42691bfe178fc9c491928ea