Add example of SoftwareConfig driving os-apply-config
Adds a minimal example of driving os-apply-config directly, e.g
with no heat-config hooks.  This is essentially how TripleO
currently uses SoftwareConfig resources, so it's useful for
reporoducing TripleO related problems.

Note, this needs rebuilding your image with the addition of the
"ntp" element, and currently the signalling won't work
https://review.openstack.org/#/c/145764/ is the update to
os-refresh-config which aims to fix that.

Change-Id: Id7a84655632da9fa6c4079954178f4abda787e4a
Related-Bug: #1389178