Create own script for running the dsvm gate
In order to fix the gate properly we need some control over the way
we setup the environment.

This patch creates a `tools\post_test_hook.sh` file that for now mirrors
the one stated in openstack-infra/project-config.

Later the project-config will be modified to use it. After that we can
start configuring the environment to our needs.

Change-Id: Ib9388e7f5001d7efa35f6f16421251c9a59ad396
Partial-Bug: #1492942