Allow ansible inventory to be configurable
This allow the user to pass a string (or path to a file) for the
ansible inventory so it is not just localhost. This is important for
things like setting up groups and possible inventory scoped variables.

Added test to ensure ansible is getting called properly using the ENV
inventory value.

Change-Id: I9440ff9dae3595a1c57b2511f03792105f1beb88