Make hooks dir discovery more flexible
Currently the hooks dir is statically set to $HEAT_CONFIG_HOOKS or
/var/lib/heat-config/hooks.

This commit changes the approach to finding the hook to execute, it
now looks for the requested hook from a list of directories. The new
directory /usr/libexec/heat-config/hooks has precedence over the old
/var/lib/heat-config/hooks.

The reason for this is that /usr/libexec is a more appropriate
location for downstream packaging to install hook scripts (RDO, debian
etc).

This approach will gracefully handle upgrading an image built with
hooks in /var/lib/heat-config/hooks followed by a package based
upgrade which switches some or all of the hooks to
/usr/libexec/heat-config/hooks.

Change-Id: Id0262e8ab1f96fa556e2ffbdf8b1b25be14ea97b