Fixed MSSQL template to support 2008, 2014, 2016 versions
It has been tested with standard 2008/2012/2014 and enterprise 2016.
I've added the functions Get-MSSQLVersion that checks for the version,
and Get-MSSQLLogFile that finds the log folder for the MSSQL version used.
The function Get-MSSQLParameters has been updated to filter some install
arguments that are incompatible with versions of MSSQL diffrent from 2012.

Change-Id: I438a74b081fa6dd38d9e39c10b11497032e070cb